#182: DiaryRestAnonTest added
